+++ etcd/include/userver/etcd/exceptions.hpp
@@ -6,7 +6,7 @@
 
 USERVER_NAMESPACE_BEGIN
 
-namespace storages::etcd {
+namespace etcd {
 
 /// @brief Base class for all etcd client exceptions
 class EtcdError : public std::runtime_error {
@@ -14,6 +14,6 @@ public:
     using std::runtime_error::runtime_error;
 };
 
-}
+}  // namespace etcd
 
 USERVER_NAMESPACE_END
