+++ etcd/src/etcd/component.cpp
@@ -1,21 +1,20 @@
-#include <userver/storages/etcd/component.hpp>
+#include <userver/etcd/component.hpp>
 
+#include <etcd/client_impl.hpp>
 #include <userver/clients/http/component.hpp>
-#include <userver/storages/etcd/settings.hpp>
+#include <userver/etcd/settings.hpp>
 #include <userver/yaml_config/merge_schemas.hpp>
-
 
 USERVER_NAMESPACE_BEGIN
 
-namespace storages::etcd {
+namespace etcd {
 
 Component::Component(const components::ComponentConfig& config, const components::ComponentContext& context)
-    :
-    ComponentBase(config, context),
-    etcd_client_ptr_(std::make_shared<impl::ClientImpl>(
-        context.FindComponent<components::HttpClient>().GetHttpClient(),
-        config.As<ClientSettings>()
-    )) {}
+    : ComponentBase(config, context),
+      etcd_client_ptr_(std::make_shared<impl::ClientImpl>(
+          context.FindComponent<components::HttpClient>().GetHttpClient(),
+          config.As<ClientSettings>()
+      )) {}
 
 yaml_config::Schema Component::GetStaticConfigSchema() {
     return yaml_config::MergeSchemas<ComponentBase>(R"(
@@ -29,22 +28,20 @@ properties:
         items:
             type: string
             description: host
-    retries:
+    attempts:
         type: integer
         description: >
-            Number of retries per one endpoints, total number of retries is number of endpoints times retries
+            Number of attempts per one endpoints, total number of attempts is number of endpoints times attempts
         minimum: 1
     request_timeout_ms:
         type: integer
-        description: Number of miliseconds to timeout request
+        description: Number of miliseconds between request attempts
         minimum: 1
 )");
 }
 
-ClientPtr Component::GetClient() {
-    return etcd_client_ptr_;
-}
+ClientPtr Component::GetClient() { return etcd_client_ptr_; }
 
-}  // namespace storages::etcd
+}  // namespace etcd
 
 USERVER_NAMESPACE_END
